vshn-lbaas-exoscale: Take security group names as inputs
We need both the security group name (to configure the LB security
group) and id (to configure the VMs) in the module. We could either take
the security group id or name as input. However, when using security
group ids as inputs, Terraform generates the following error for an
empty state:

```
Error: Null value found in list

  with module.lb.exoscale_security_group_rules.load_balancers,
  on .terraform/modules/lb/modules/vshn-lbaas-exoscale/security_groups.tf line 11, in resource "exoscale_security_group_rules" "load_balancers":
  11: resource "exoscale_security_group_rules" "load_balancers" {

Null values are not allowed for this attribute value.
```

So we use the security group name as input and use a datasource to fetch
the id for the security group.